+++ app/src/main/java/com/vag/lmsapp/app/packages/PackageItem.kt
@@ -9,6 +9,6 @@ data class PackageItem(
     val serviceId: UUID,
     val name: String,
     val price: Float,
-    val quantity: Float,
+    val quantity: Int,
     var deleted: Boolean,
 ) : Parcelable
